doc: just use markdown for plain text
