+++ pretty/src/main/java/org/aya/pretty/backend/md/DocMdPrinter.java
@@ -5,6 +5,7 @@ package org.aya.pretty.backend.md;
 import org.aya.pretty.backend.html.DocHtmlPrinter;
 import org.aya.pretty.backend.html.HtmlConstants;
 import org.aya.pretty.backend.string.Cursor;
+import org.aya.pretty.backend.string.StringPrinter;
 import org.aya.pretty.doc.Doc;
 import org.jetbrains.annotations.NotNull;
 
@@ -64,11 +65,8 @@ public class DocMdPrinter extends DocHtmlPrinter<DocMdPrinter.Config> {
   }
 
   @Override protected void renderHardLineBreak(@NotNull Cursor cursor, EnumSet<Outer> outer) {
-    if (outer.contains(Outer.List)) {
-      cursor.lineBreakWith("<br/>\n");
-    } else {
-      cursor.lineBreakWith("\n");
-    }
+    if (outer.contains(Outer.List)) cursor.lineBreakWith("<br/>\n");
+    else cursor.lineBreakWith("\n");
   }
 
   @Override
@@ -100,34 +98,8 @@ public class DocMdPrinter extends DocHtmlPrinter<DocMdPrinter.Config> {
     });
   }
 
-  /**
-   * @see org.aya.pretty.backend.string.StringPrinter#renderList
-   */
   @Override protected void renderList(@NotNull Cursor cursor, @NotNull Doc.List list, EnumSet<Outer> outer) {
-    var isTopLevel = !outer.contains(Outer.List);
-    requireLineStart(cursor);
-    if (isTopLevel) renderHardLineBreak(cursor, outer);
-
-    list.items().forEachIndexed((idx, item) -> {
-      requireLineStart(cursor);
-
-      var pre = list.isOrdered()
-        ? (idx + 1) + "."
-        : "+";
-      var content = Doc.nest(pre.length() + 1, item);
-
-      // render pre
-      cursor.visibleContent(pre);
-      cursor.visibleContent(" ");
-
-      // render
-      renderDoc(cursor, content, EnumSet.of(Outer.List));
-    });
-
-    if (isTopLevel) {
-      requireLineStart(cursor);
-      renderHardLineBreak(cursor, outer);
-    }
+    StringPrinter.renderList(this, cursor, list, outer);
   }
 
   @Override protected void renderCodeBlock(@NotNull Cursor cursor, @NotNull Doc.CodeBlock block, EnumSet<Outer> outer) {
