+++ pretty/src/test/java/org/aya/pretty/DocStringPrinterTest.java
@@ -122,11 +122,11 @@ public class DocStringPrinterTest {
     assertEquals("""
       what
             
-      * fir
+      + fir
         st
-      * second
-        * second.1
-      * third
+      + second
+        + second.1
+      + third
             
       """, list.commonRender());
   }
